# inventory-api/docs/openapi/runtime-baseline.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "inventory-api runtime baseline",
    "version": "0.5.0",
    "description": "Baseline OpenAPI parcial y factual para superficies críticas verificadas en runtime. No cubre todo el monolito y se complementa con un manifiesto explícito de exclusiones contractuales."
  },
...
  "servers": [
    {
      "url": "/"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/api/orders": {
      "get": {
        "operationId": "listOrders",
        "summary": "Listar pedidos visibles para roles operativos habilitados",
        "tags": [
          "orders"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-runtime-source": {
          "mountPath": "/api/orders",
          "routeFile": "src/routes/order.routes.js",
          "expressPath": "/"
        },
        "responses": {
          "200": {
            "description": "Listado de pedidos"
          },
          "403": {
            "description": "Sin permisos"
          }
        }
      },
      "post": {
        "operationId": "createOrder",
        "summary": "Registrar pedido de ventas",
        "tags": [
          "orders"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-runtime-source": {
          "mountPath": "/api/orders",
          "routeFile": "src/routes/order.routes.js",
          "expressPath": "/"
        },
        "responses": {
          "201": {
            "description": "Pedido creado"
          },
          "403": {
            "description": "Sin permisos"
          }
        }
      }
    },
...
    "/api/production/orders": {
      "get": {
        "operationId": "listProductionOrders",
        "summary": "Listar órdenes de producción visibles para la compañía",
        "tags": [
          "production"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-governance": {
          "routePolicy": "production.view",
          "actorScope": "company-admin"
        },
        "x-runtime-source": {
          "mountPath": "/api/production",
          "routeFile": "src/routes/production.routes.js",
          "expressPath": "/orders"
        },
        "responses": {
          "200": {
            "description": "Listado de órdenes de producción"
          },
          "403": {
            "description": "Sin permisos"
          }
        }
      },
      "post": {
        "operationId": "createProductionOrder",
        "summary": "Crear orden de producción",
        "tags": [
          "production"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-governance": {
          "routePolicy": "production.create",
          "actorScope": "company-admin"
        },
        "x-runtime-source": {
          "mountPath": "/api/production",
          "routeFile": "src/routes/production.routes.js",
          "expressPath": "/orders"
        },
        "responses": {
          "201": {
            "description": "Orden de producción creada"
          },
          "400": {
            "description": "Payload inválido"
          },
          "403": {
            "description": "Sin permisos"
          }
        }
      }
    },
    "/api/production/orders/{id}": {
      "get": {
        "operationId": "getProductionOrder",
        "summary": "Obtener detalle de orden de producción",
        "tags": [
          "production"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-governance": {
          "routePolicy": "production.view",
          "actorScope": "company-admin"
        },
        "x-runtime-source": {
          "mountPath": "/api/production",
          "routeFile": "src/routes/production.routes.js",
          "expressPath": "/orders/:id"
        },
        "responses": {
          "200": {
            "description": "Detalle de orden de producción"
          },
          "403": {
            "description": "Sin permisos"
          },
          "404": {
            "description": "Orden no encontrada"
          }
        }
      }
    },
    "/api/production/orders/{id}/submit": {
      "post": {
        "operationId": "submitProductionOrder",
        "summary": "Enviar orden de producción a aprobación",
        "tags": [
          "production"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-governance": {
          "routePolicy": "production.create",
          "actorScope": "company-admin"
        },
        "x-runtime-source": {
          "mountPath": "/api/production",
          "routeFile": "src/routes/production.routes.js",
          "expressPath": "/orders/:id/submit"
        },
        "responses": {
          "200": {
            "description": "Orden enviada a aprobación"
          },
          "403": {
            "description": "Sin permisos"
          },
          "404": {
            "description": "Orden no encontrada"
          },
          "409": {
            "description": "Transición inválida"
          }
        }
      }
    },
    "/api/production/orders/{id}/approve": {
      "post": {
        "operationId": "approveProductionOrder",
        "summary": "Aprobar orden de producción",
        "tags": [
          "production"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-governance": {
          "routePolicy": "production.approve",
          "actorScope": "company-admin"
        },
        "x-runtime-source": {
          "mountPath": "/api/production",
          "routeFile": "src/routes/production.routes.js",
          "expressPath": "/orders/:id/approve"
        },
        "responses": {
          "200": {
            "description": "Orden aprobada"
          },
          "400": {
            "description": "Payload inválido"
          },
          "403": {
            "description": "Sin permisos"
          },
          "404": {
            "description": "Orden no encontrada"
          },
          "409": {
            "description": "Transición inválida"
          }
        }
      }
    },
    "/api/production/orders/{id}/start": {
      "post": {
        "operationId": "startProductionOrder",
        "summary": "Iniciar orden de producción aprobada",
        "tags": [
          "production"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-governance": {
          "routePolicy": "production.execute",
          "actorScope": "company-admin"
        },
        "x-runtime-source": {
          "mountPath": "/api/production",
          "routeFile": "src/routes/production.routes.js",
          "expressPath": "/orders/:id/start"
        },
        "responses": {
          "200": {
            "description": "Orden iniciada"
          },
          "403": {
            "description": "Sin permisos"
          },
          "404": {
            "description": "Orden no encontrada"
          },
          "409": {
            "description": "Transición inválida"
          }
        }
      }
    },
    "/api/production/orders/{id}/complete": {
      "post": {
        "operationId": "completeProductionOrder",
        "summary": "Completar orden de producción",
        "tags": [
          "production"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-governance": {
          "routePolicy": "production.complete",
          "actorScope": "company-admin"
        },
        "x-runtime-source": {
          "mountPath": "/api/production",
          "routeFile": "src/routes/production.routes.js",
          "expressPath": "/orders/:id/complete"
        },
        "responses": {
          "200": {
            "description": "Orden completada"
          },
          "400": {
            "description": "Payload inválido"
          },
          "403": {
            "description": "Sin permisos"
          },
          "404": {
            "description": "Orden no encontrada"
          },
          "409": {
            "description": "QA pendiente o transición inválida"
          }
        }
      }
    },
    "/api/production/orders/{id}/cancel": {
      "post": {
        "operationId": "cancelProductionOrder",
        "summary": "Cancelar orden de producción",
        "tags": [
          "production"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-governance": {
          "routePolicy": "production.cancel",
          "actorScope": "company-admin"
        },
        "x-runtime-source": {
          "mountPath": "/api/production",
          "routeFile": "src/routes/production.routes.js",
          "expressPath": "/orders/:id/cancel"
        },
        "responses": {
          "200": {
            "description": "Orden cancelada"
          },
          "403": {
            "description": "Sin permisos"
          },
          "404": {
            "description": "Orden no encontrada"
          },
          "409": {
            "description": "Transición inválida"
          }
        }
      }
    },
...
  }
}